cleanup fix

+++ utils.py
@@ -1,13 +1,17 @@
 from glob import glob
 from os import remove
+from os.path import isdir
 from shutil import rmtree
 
 
-def cleanup():
-    patterns_to_delete = ["calc.fc", "calc.pvd", "cubit*", "outs.csv"]
+def cleanup(start):
+    patterns_to_delete = ["calc.fc", "calc.pvd", "cubit*"]
     to_delete = []
     for pattern in patterns_to_delete:
         to_delete.extend(glob(pattern))
     for item in to_delete:
         remove(item)
-    rmtree("calc")
+    if isdir("calc"):
+        rmtree("calc")
+    if start and isdir("outs.csv"):
+        remove("outs.csv")
